app/requirements.txt
aiodns==2.0.0
aiohttp==3.7.3
aiohttp-socks==0.5.5
async-timeout==3.0.1
awscli==1.18.207
bleach==3.1.1
botocore==1.19.47
cachetools==4.2.1
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
dataclasses==0.6
docutils==0.15.2
elasticsearch==7.10.1
fake-useragent==0.1.11
Flask==1.1.2
geographiclib==1.50
geopy==2.1.0
google-api-core==1.26.0
google-auth==1.26.1
google-cloud-core==1.6.0
google-cloud-storage==1.36.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
googletransx==2.4.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
keyring==21.1.0
MarkupSafe==1.1.1
multidict==5.1.0
packaging==20.9
pkginfo==1.5.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
python-socks==1.2.0
pytz==2021.1
PyYAML==5.3.1
readme-renderer==24.0
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.5
s3transfer==0.3.3
schedule==1.0.0
six==1.15.0
twine==3.1.1
typing-extensions==3.7.4.3
urllib3==1.26.3
webencodings==0.5.1
Werkzeug==1.0.1
yarl==1.6.3
